+++ core/singletons/template.js
@@ -1,0 +1,13 @@
+export default class SingletonClass {
+	destroy () {
+		throw new Error("Method not implemented.");
+	}
+
+	static singleton () {
+		throw new Error("Method not implemented.");
+	}
+
+	get modulePath () {
+		throw new Error("Method not implemented.");
+	}
+}